Scraped business logic, remodeled data layer structure
