+++ DirectTorrent.Data.Yify.ApiWrappers/DirectTorrent.Data.Yify.ApiWrappers.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{27727704-EA27-496F-BBB9-7B1878AA4C05}</ProjectGuid>
+    <ProjectGuid>{8CCF7C6E-42CB-4646-A22B-0C5E42D11E17}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>DirectTorrent.Logic.DTO</RootNamespace>
-    <AssemblyName>DirectTorrent.Logic.DTO</AssemblyName>
+    <RootNamespace>DirectTorrent.Data.Yify.ApiWrappers</RootNamespace>
+    <AssemblyName>DirectTorrent.Data.Yify.ApiWrappers</AssemblyName>
     <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -30,6 +30,10 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="Newtonsoft.Json, Version=6.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\packages\Newtonsoft.Json.6.0.8\lib\net40\Newtonsoft.Json.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -39,8 +43,18 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Class1.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ApiResponse.cs" />
+    <Compile Include="ApiWrapper.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\DirectTorrent.Data.Models\DirectTorrent.Data.Yify.Models.csproj">
+      <Project>{cf058cc9-4a20-4cbc-a25f-48b01e79d828}</Project>
+      <Name>DirectTorrent.Data.Yify.Models</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="packages.config" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
