+++ DirectTorrent.Data.Yify.Models/DirectTorrent.Data.Yify.Models.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{8CCF7C6E-42CB-4646-A22B-0C5E42D11E17}</ProjectGuid>
+    <ProjectGuid>{CF058CC9-4A20-4CBC-A25F-48B01E79D828}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>DirectTorrent.Data.ApiWrappers</RootNamespace>
-    <AssemblyName>DirectTorrent.Data.ApiWrappers</AssemblyName>
+    <RootNamespace>DirectTorrent.Data.Yify.Models</RootNamespace>
+    <AssemblyName>DirectTorrent.Data.Yify.Models</AssemblyName>
     <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -43,10 +43,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ApiData.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Yify\ApiData.cs" />
-    <Compile Include="Yify\ApiResponse.cs" />
-    <Compile Include="Yify\ApiWrapper.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
