+++ DirectTorrent.Logic.Services/MovieRepository.cs
@@ -3,9 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
+using DirectTorrent.Logic.Models;
+
 namespace DirectTorrent.Logic.Services
 {
     public static class MovieRepository
     {
+        public List<Movie> 
     }
 }
